Do not use hijack for obfuscating secrets
